`size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: /media.
This CL makes use of the new `size_t`-based APIs in the remaining
directories like `//components`, `//extensions`, or `//storage`.
See https://crrev.com/c/5414148 for motivation.

This CL may have introduced new integer conversions - in those cases the
CL attempts to have well-defined behavior (e.g. using
`base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
integer operations (e.g. old `int` => `uint32_t` / `size_t` conversions,
or old additions that may potentially overflow).

This CL was uploaded by git cl split.

R=tguilbert@chromium.org

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5461341
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288918}

+++ media/mojo/common/mojo_data_pipe_read_write.h
@@ -54,13 +54,13 @@ class MojoDataPipeReader {
   raw_ptr<uint8_t, AllowPtrArithmetic> current_buffer_ = nullptr;
 
   // The number of bytes to be read for the current read request.
-  uint32_t current_buffer_size_ = 0;
+  size_t current_buffer_size_ = 0;
 
   // The current once callback to be called when read completes.
   DoneCB done_cb_;
 
   // Number of bytes already read into the current buffer.
-  uint32_t bytes_read_ = 0;
+  size_t bytes_read_ = 0;
 };
 
 // Write a certain amount of data into a mojo data pipe by request.
@@ -111,7 +111,7 @@ class MojoDataPipeWriter {
   DoneCB done_cb_;
 
   // Number of bytes already written from the current buffer.
-  uint32_t bytes_written_ = 0;
+  size_t bytes_written_ = 0;
 };
 
 }  // namespace media
